% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nuts.R
\name{.assemble_output_all}
\alias{.assemble_output_all}
\title{Assemble all-levels output (0..3) with status from level 3}
\usage{
.assemble_output_all(lat, lon, out0, out1, out2, out3, year, resolution)
}
\description{
Assemble all-levels output (0..3) with status from level 3
}
\keyword{internal}
