% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nuts.R
\name{.assemble_output_single}
\alias{.assemble_output_single}
\title{Assemble single-level output}
\usage{
.assemble_output_single(lat, lon, joined, level, year, resolution)
}
\description{
Assemble single-level output
}
\keyword{internal}
