% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nuts.R
\name{.coords_to_sf}
\alias{.coords_to_sf}
\title{Convert numeric lat/lon vectors to sf POINT (WGS84)}
\usage{
.coords_to_sf(latitude, longitude, crs_input = 4326)
}
\description{
Convert numeric lat/lon vectors to sf POINT (WGS84)
}
\keyword{internal}
