% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nuts.R
\name{.filter_country}
\alias{.filter_country}
\title{Country filter by CNTR_CODE (if available)}
\usage{
.filter_country(shp, country)
}
\description{
Country filter by CNTR_CODE (if available)
}
\keyword{internal}
