% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nuts.R
\name{.get_cached_shape}
\alias{.get_cached_shape}
\title{Cached Eurostat geospatial layer}
\usage{
.get_cached_shape(level, year, resolution, verbose = TRUE)
}
\description{
Cached Eurostat geospatial layer
}
\keyword{internal}
