% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nuts.R
\name{.join_points_to_nuts}
\alias{.join_points_to_nuts}
\title{Spatial join + optional nearest fallback (robust; km units)}
\usage{
.join_points_to_nuts(
  pts,
  shp,
  match_strategy = "within",
  nearest_max_km = Inf,
  level = NA_integer_
)
}
\description{
Spatial join + optional nearest fallback (robust; km units)
}
\keyword{internal}
