% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nuts.R
\name{.prefilter_by_bbox}
\alias{.prefilter_by_bbox}
\title{Pre-filter polygons by bounding box of points (safe with fallback)}
\usage{
.prefilter_by_bbox(shp, pts)
}
\description{
Pre-filter polygons by bounding box of points (safe with fallback)
}
\keyword{internal}
