% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nuts.R
\name{.render_map}
\alias{.render_map}
\title{Render the map (ggplot2)}
\usage{
.render_map(md, border_col, low_col, high_col, id_col, uid_col, show_points)
}
\description{
Render the map (ggplot2)
}
\keyword{internal}
