% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nuts.R
\name{.standardize_nuts_input}
\alias{.standardize_nuts_input}
\title{Normalise get_nuts() output for mapping}
\usage{
.standardize_nuts_input(nuts, map_level)
}
\description{
Normalise get_nuts() output for mapping
}
\keyword{internal}
