% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nuts.R
\name{.validate_get_nuts_params}
\alias{.validate_get_nuts_params}
\title{Parameter validation for get_nuts()}
\usage{
.validate_get_nuts_params(
  latitude,
  longitude,
  level,
  year,
  resolution,
  crs,
  country,
  match_strategy,
  nearest_max_km
)
}
\description{
Parameter validation for get_nuts()
}
\keyword{internal}
